Introduce `metrics "pipeline"`
This PR adds `metrics "pipeline"` as an easier-to-understand version of
the now deprecated `metrics "operator"`.

Operator metrics were very detailed. We simply generated too many of
them, causing them to take away significant compute resources. Looking
at how we utilize operator metrics in the Tenzir Platform, it became
clear that we just need a small subset of the information that they
provided. The new pipeline metrics are exactly that subset, showing the
ingress and egress of every managed pipeline pre-aggregated in windows
of ten seconds.

We will remove operator metrics once we have adapated the Tenzir
Platform to make use of pipeline metrics instead.